fix minor bug

+++ Fibonacci/app/src/main/java/cytech/android/fibonacci/Java.java
@@ -1,7 +1,7 @@
 package cytech.android.fibonacci;
 
 public class Java {
-    public static int Fibonacci(int i) {
+    public static long Fibonacci(int i) {
         if ((i == 0) || (i == 1))
             return i;
         else
