+++ Fibonacci/app/src/main/java/cytech/android/fibonacci/Kotlin.kt
@@ -1,5 +1,5 @@
 package cytech.android.fibonacci
 
-fun Fibonacci(i: Int): Int {
-    return if (i == 0 || i == 1) i else Fibonacci(i - 1) + Fibonacci(i - 2)
+fun Fibonacci(i: Int): Long {
+    return if (i == 0 || i == 1) i.toLong() else Fibonacci(i - 1) + Fibonacci(i - 2)
 }
